+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Compatibility:
 * Define `#getch` and `#getpass` on `StringIO` when `io/console` is required.
 * Implemented `rb_uv_to_utf8` (#1998).
 * Single character IDs now behave more like those in MRI to improve C extension compatibility, so `rb_funcall(a, '+', b)` will now do the same thing as in MRI.
+* Removed extra public methods on `String`.
 
 Performance: